+++ packages/deb/make-yuneta-agent-deb.sh
@@ -670,7 +670,7 @@ Priority: optional
 Architecture: ${ARCHITECTURE}
 Homepage: https://yuneta.io
 Maintainer: ArtGins S.L. <support@artgins.com>
-Depends: adduser, lsb-base, rsync, locales, rsyslog, init-system-helpers
+Depends: adduser, lsb-base, rsync, locales, rsyslog, init-system-helpers, gdb
 Recommends: curl, vim, sudo, tree, pipx, fail2ban, net-tools, locate
 Suggests: git, mercurial, make, cmake, ninja-build, gcc, clang, g++, python3-dev, python3-pip, python3-setuptools, python3-tk, python3-wheel, python3-venv, libjansson-dev, libpcre2-dev, liburing-dev, libcurl4-openssl-dev, zlib1g-dev, libssl-dev, perl, dos2unix, postgresql-server-dev-all, libpq-dev, kconfig-frontends, telnet, patch, gettext, snapd
 Description: Yuneta's Agent